// srv/assets/ts/apg-gui/lib/enums/ApgGui_eInputType.ts
/** -----------------------------------------------------------------------
 * @module [apg-gui]
 * @author [APG] ANGELI Paolo Giusto
 * @version 0.0.1 [APG 2023/08/09]
 * @version 0.0.8 [APG 2023/11/12]
 * -----------------------------------------------------------------------
 */

/**
 * Types of input elements
 */
export enum ApgGui_eInputType {
    BUTTON = "button",
    CHECK_BOX = "checkbox",
    COLOR = "color",
    DATE = "date",
    DATETIME_LOCAL = "datetime-local",
    EMAIL = "email",
    FILE = "file",
    HIDDEN = "hidden",
    IMAGE = "image",
    MONTH = "month",
    NUMBER = "number",
    PASSWORD = "password",
    RADIO = "radio",
    RANGE = "range",
    RESET = "reset",
    SEARCH = "search",
    SUBMIT = "submit",
    TEL = "tel",
    TEXT = "text",
    TIME = "time",
    URL = "url",
    WEEK = "week"
}