// srv/assets/ts/apg-gui/lib/enums/ApgGui_eFormElementType.ts
/** -----------------------------------------------------------------------
 * @module [apg-gui]
 * @author [APG] ANGELI Paolo Giusto
 * @version 0.0.1 [APG 2023/08/09]
 * @version 0.0.8 [APG 2023/11/12]
 * -----------------------------------------------------------------------
 */

/**
 * Types of elements for forms and dialogs
 */
export enum ApgGui_eFormElementType {

    PARAGRAPH = 'paragraph',
    DIV = 'div',
    DETAILS = 'details',
    INPUT = "input",
    LABEL = "label",
    SELECT = "select",
    TEXT_AREA = "textarea",
    BUTTON = "button",
    FIELD_SET = "fieldset",
    LEGEND = "legend",
    DATA_LIST = "datalist",
    OUTPUT = "output",
    OPTION = "option",
    OPT_GROUP = "optgroup",
    DIALOG = "dialog"

}